Begin mockup of PPU
When a read to the PPU is done, PPU flags need to change and data needs to be modified.
This means we have a problem with mutability if the `read8()` function is to stay non mutable.
This also implements all of the PPU registers as structs. In theory we can maybe `impl` the structs
and modify the data at that point?

The main goal here is to get NMIs firing and vblank-ing.