Render pattern tables with test palette
This renders the two pattern tables with a hacky palette implementation.
We use a `u32` palette in this case to give to our frame buffer.
We will need to support 8 bit palette lookup (i.e use the NES palette),
but to do so we need to convert the values back to u32 to be rendered.